[TS-286] Refactor: Star 컴포넌트 삭제
자주 사용하지 않아서 삭제하고 Img 컴포넌트와 병합함
